+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - move device types to defaults for SIG credentials feature template
 - rename `traffic_class` to `traffic_classes` in both transport and service IPv6 ACL match actions
+- add support for policy object Security Data IP Prefix
 
 ## 1.3.0